+++ CHANGELOG.md
@@ -1,3 +1,10 @@
-# 0.1.0
+## v0.2.0
+#### Added
 
-* Initial release.
+* Add support several colors in one property
+* Fix regexp for different color types
+* Add support color keywords
+* Add opportunity to set custom selector, which will contain variables
+
+## v0.1.0
+Initial release